feat: Implemented cdRefactor accessory dialogs and implement global delete dialog component
- Replaced individual delete confirmation dialogs in AccessoriesPage, MaterialsPage, and ContactsPage with a unified GlobalDeleteDialog component.
- Introduced AddAccessoryDialog and EditAccessoryDialog components for adding and editing accessories, respectively, with improved state management and validation.
- Enhanced the AccessoriesPage to utilize async components for better performance and loading states.
- Cleaned up unused code and improved overall structure for better maintainability.
