+++ Frontend/src/components/Cabs/DeleteDialog.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import { computed } from 'vue';
+import GlobalDeleteDialog from 'src/components/Global/DeleteDialog.vue';
 
-const props = defineProps({
+defineProps({
     modelValue: {
         type: Boolean,
         required: true,
@@ -14,49 +14,30 @@ const props = defineProps({
         type: String,
         default: 'this item',
     },
+    confirmText: {
+        type: String,
+        default: 'Delete',
+    },
+    cancelText: {
+        type: String,
+        default: 'Cancel',
+    },
+    iconName: {
+        type: String, 
+        default: 'warning',
+    }
 });
 
 const emit = defineEmits<{
     (e: 'update:modelValue', value: boolean): void
     (e: 'confirm-delete'): void
 }>();
-
-const capitalizedItemType = computed(() => {
-    return props.itemType.charAt(0).toUpperCase() + props.itemType.slice(1);
-});
-
-function confirm() {
-    emit('confirm-delete');
-    closeDialog(); // Usually close after confirm
-}
-
-function closeDialog() {
-    emit('update:modelValue', false);
-}
-
 </script>
 
 <template>
-    <q-dialog :model-value="modelValue" persistent @update:model-value="closeDialog">
-        <q-card>
-            <q-card-section class="row items-center">
-                <q-avatar icon="warning" color="negative" text-color="white" />
-                <span class="q-ml-sm text-h6">Delete {{ capitalizedItemType }}</span>
-            </q-card-section>
-
-            <q-card-section>
-                Are you sure you want to delete <span class="text-weight-bold">{{ itemName }}</span>? This action cannot
-                be undone.
-            </q-card-section>
-
-            <q-card-actions align="right" class="q-pa-md">
-                <q-btn flat label="Cancel" @click="closeDialog" />
-                <q-btn flat label="Delete" color="negative" @click="confirm" />
-            </q-card-actions>
-        </q-card>
-    </q-dialog>
+    <GlobalDeleteDialog
+        v-bind="$props"
+        @update:modelValue="(val) => emit('update:modelValue', val)"
+        @confirm-delete="() => emit('confirm-delete')"
+    />
 </template>
-
-<style scoped lang="sass">
-/* Add any specific styles if needed */
-</style>
